Move Feature to models subpackage
